% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_codes.R
\name{get_codes}
\alias{get_codes}
\title{Download ISTAT codes for italian regions, provinces and municipalities.}
\usage{
get_codes(type)
}
\arguments{
\item{type}{character. The argument can be set to \code{region}, \code{province}, or \code{municipality} according to which codes are needed.}
}
\value{
a data.frame object
}
\description{
get_codes allows to retrieve codes for regions, provinces, and municipality, filtering for the type of codes needed.
}
\examples{
data <- get_codes("region")

}
\author{
Lorena Ricciotti
}
