% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_type_RENDIS.R
\name{get_type_RENDIS}
\alias{get_type_RENDIS}
\title{Retrieve information about soil defense type of the ReNDiS database}
\usage{
get_type_RENDIS()
}
\value{
Return an object of class \code{data.frame}

Types:

\itemize{
\item{Frana} \strong{=>} Landslide
\item{Non definito} \strong{=>} Not defined
\item{Alluvione} \strong{=>} Flooding
\item{Misto} \strong{=>} Mixed
\item{Valanga} \strong{=>} Avalanche
\item{Incendio} \strong{=>} Wildfire
\item{Costiero} \strong{=>} Coastal
}
}
\description{
The \code{get_type_RENDIS} function returns the list of type of interventions for soil defense contained in the ReNDiS database.
}
\examples{
get_type_RENDIS()

}
\references{
\href{http://www.rendis.isprambiente.it/rendisweb/}{ReNDiS}
}
\author{
Lorena Ricciotti
}
