% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funds_bar.R
\name{plot_funds_bar}
\alias{plot_funds_bar}
\title{Repartition of Financial Funds Allocation: Investment Amounts Barplot}
\usage{
plot_funds_bar(data, var_col)
}
\arguments{
\item{data}{Dataset of class 'data.frame'. Specify the dataset from which to take information.}

\item{var_col}{integer value. Specify the number of the columns associated with the variable to visualize.}
}
\value{
An object of class \code{gg} and \code{ggplot} representing the barplot
}
\description{
The \code{plot_funds_bar} function creates a barplot to visually represent the distribution of financial funds allocation across different investment channels.
}
\examples{
data(OCpoint)
plot_funds_bar(OCpoint, var_col = c(10:15))
#Barplot visualizing the total amount allocated by each fund.

}
\author{
Lorena Ricciotti
}
